Close #472: [`test-refined4s-core-without-cats`] Run `core` tests for `refined4s.types.all` to ensure the `core` doesn't require `cats`
