+++ modules/test-refined4s-core-without-cats/native/src/test/scala/refined4s/types/networkCompatSpec.scala
@@ -1,10 +1,10 @@
 package refined4s.types
 
-import cats.syntax.all.*
 import hedgehog.*
 import hedgehog.runner.*
+import refined4s.predef4testing.*
 
-import java.net.{URI, URL}
+import java.net.URI
 
 /** @author Kevin Lee
   * @since 2024-09-05
@@ -15,10 +15,9 @@ trait networkCompatSpec {
 
   def compatTests: List[Test] = List(
     property("test Uri.toUrl", testUriToUrl),
-    property("test Uri.toURL", testUriToURL),
     //
     example("test Url(valid URL String)", testUrlApply),
-    example("test Url(URL)", testUrlApplyURL),
+    //    example("test Url(URL)", testUrlApplyURL),
     example("test Url(invalid URL String)", testUrlApplyInvalid),
     property("test Url.from(valid)", testUrlFromValid),
     property("test Url.from(invalid)", testUrlFromInvalid),
@@ -26,7 +25,7 @@ trait networkCompatSpec {
     property("test Url.unsafeFrom(invalid)", testUrlUnsafeFromInvalid),
     property("test Url.value", testUrlValue),
     property("test Url.unapply", testUrlUnapply),
-    property("test Url.toURL", testUrlToURL),
+    //    property("test Url.toURL", testUrlToURL),
     property("test Url.toUri", testUrlToUri),
     property("test Url.toURI", testUrlToURI),
     example("test network.isValidateUrl(valid URL String)", testNetworkIsValidateUrlValid),
@@ -45,43 +44,32 @@ trait networkCompatSpec {
       actual ==== expected
     }
 
-  def testUriToURL: Property =
-    for {
-      uri <- networkGens.genUrlString.log("uri")
-    } yield {
-      @SuppressWarnings(Array("org.wartremover.warts.JavaNetURLConstructors"))
-      val expected = new URL(uri)
-      val actual   = Uri.unsafeFrom(uri).toURL
-
-      actual ==== expected
-    }
-
   //
 
   def testUrlApply: Result = {
     @SuppressWarnings(Array("org.wartremover.warts.JavaNetURLConstructors"))
-    val expected = new URL("https://github.com/kevin-lee/refined4s")
+    val expected = URI("https://github.com/kevin-lee/refined4s")
     val actual   = Url("https://github.com/kevin-lee/refined4s")
     Result.all(
       List(
         actual.value ==== expected.toString,
-        actual.toURL ==== expected,
+        actual.toURI ==== expected,
       )
     )
   }
 
-  def testUrlApplyURL: Result = {
-    @SuppressWarnings(Array("org.wartremover.warts.JavaNetURLConstructors"))
-    val url      = new URL("https://github.com/kevin-lee/refined4s")
-    val expected = url
-    val actual   = Url(url)
-    Result.all(
-      List(
-        actual.value ==== expected.toString,
-        actual.toURL ==== expected,
-      )
-    )
-  }
+  //  def testUrlApplyURL: Result = {
+  //    @SuppressWarnings(Array("org.wartremover.warts.JavaNetURLConstructors"))
+  //    val url      = new URL("https://github.com/kevin-lee/refined4s")
+  //    val expected = url
+  //    val actual   = Url(url)
+  //    Result.all(
+  //      List(
+  //        actual.value ==== expected.toString,
+  //        actual.toURL ==== expected,
+  //      )
+  //    )
+  //  }
 
   def testUrlApplyInvalid: Result = {
     import scala.compiletime.testing.typeChecks
@@ -110,19 +98,15 @@ trait networkCompatSpec {
 
   def testUrlFromValid: Property =
     for {
-      uri <- networkGens.genUrlString.log("uri")
+      url <- networkGens.genUrlString.log("url")
     } yield {
-      val expected = Url.unsafeFrom(uri).asRight
-      val actual   = Url.from(uri)
+      val expected: Either[String, Url] = Right(Url.unsafeFrom(url))
 
-      @SuppressWarnings(Array("org.wartremover.warts.JavaNetURLConstructors"))
-      val expectedUrl = new URL(uri).asRight
-      val actualUrl   = actual.map(_.toURL)
+      val actual = Url.from(url)
 
       Result.all(
         List(
-          actual ==== expected,
-          actualUrl ==== expectedUrl,
+          actual ==== expected
         )
       )
     }
@@ -154,19 +138,15 @@ trait networkCompatSpec {
 
   def testUrlUnsafeFromValid: Property =
     for {
-      uri <- networkGens.genUrlString.log("uri")
+      url <- networkGens.genUrlString.log("url")
     } yield {
-      val expected = Url.from(uri)
-      val actual   = Url.unsafeFrom(uri)
-
-      @SuppressWarnings(Array("org.wartremover.warts.JavaNetURLConstructors"))
-      val expectedUrl = new URL(uri)
-      val actualUrl   = actual.toURL
+      val expected = Url.from(url)
+      val actual   = Url.unsafeFrom(url)
 
       Result.all(
         List(
           actual.asRight ==== expected,
-          actualUrl ==== expectedUrl,
+          expected.fold(err => Result.failure.log(s"expected Url.from(url) results in error: $err"), actual ==== _),
         )
       )
     }
@@ -226,16 +206,16 @@ trait networkCompatSpec {
       }
     }
 
-  def testUrlToURL: Property =
-    for {
-      uri <- networkGens.genUrlString.log("uri")
-    } yield {
-      @SuppressWarnings(Array("org.wartremover.warts.JavaNetURLConstructors"))
-      val expected = new URL(uri)
-      val actual   = Url.unsafeFrom(uri).toURL
-
-      actual ==== expected
-    }
+  //  def testUrlToURL: Property =
+  //    for {
+  //      uri <- networkGens.genUrlString.log("uri")
+  //    } yield {
+  //      @SuppressWarnings(Array("org.wartremover.warts.JavaNetURLConstructors"))
+  //      val expected = new URL(uri)
+  //      val actual   = Url.unsafeFrom(uri).toURL
+  //
+  //      actual ==== expected
+  //    }
 
   def testUrlToUri: Property =
     for {
