+++ modules/test-refined4s-core-without-cats/shared/src/test/scala/refined4s/types/allSpec.scala
@@ -1,0 +1,11 @@
+package refined4s.types
+
+import hedgehog.*
+import hedgehog.runner.*
+
+/** @author Kevin Lee
+  * @since 2025-08-26
+  */
+object allSpec extends Properties {
+  override def tests: List[Test] = networkSpec.tests ++ stringsSpec.tests ++ numericSpec.tests
+}
