+++ modules/test-refined4s-core-without-cats/shared/src/test/scala/refined4s/types/networkGens.scala
